No circular deps
